+++ user-interface/src/main/java/life/qbic/datamanager/files/structure/ExampleProvider.java
@@ -1,6 +1,4 @@
-package life.qbic.datamanager.importing.parser;
-
-import life.qbic.datamanager.files.structure.Column;
+package life.qbic.datamanager.files.structure;
 
 /**
  * TODO!
