+++ user-interface/src/main/java/life/qbic/datamanager/files/structure/measurement/NGSMeasurementEditColumn.java
@@ -3,8 +3,8 @@ package life.qbic.datamanager.files.structure.measurement;
 import java.util.Arrays;
 import java.util.Optional;
 import life.qbic.datamanager.files.structure.Column;
-import life.qbic.datamanager.importing.parser.ExampleProvider;
-import life.qbic.datamanager.importing.parser.ExampleProvider.Helper;
+import life.qbic.datamanager.files.structure.ExampleProvider;
+import life.qbic.datamanager.files.structure.ExampleProvider.Helper;
 
 /**
  * <b>NGS Measurement Columns</b>
