+++ user-interface/src/main/java/life/qbic/datamanager/files/structure/sample/ConfoundingVariableColumn.java
@@ -2,7 +2,7 @@ package life.qbic.datamanager.files.structure.sample;
 
 import java.util.Optional;
 import life.qbic.datamanager.files.structure.Column;
-import life.qbic.datamanager.importing.parser.ExampleProvider.Helper;
+import life.qbic.datamanager.files.structure.ExampleProvider.Helper;
 import life.qbic.projectmanagement.application.confounding.ConfoundingVariableService.VariableReference;
 
 public record ConfoundingVariableColumn(VariableReference variableReference, int index,
