+++ user-interface/src/main/java/life/qbic/datamanager/views/projects/project/samples/registration/batch/EditSampleBatchDialog.java
@@ -26,11 +26,11 @@ import life.qbic.application.commons.ApplicationException;
 import life.qbic.datamanager.download.DownloadContentProvider.XLSXDownloadContentProvider;
 import life.qbic.datamanager.download.DownloadProvider;
 import life.qbic.datamanager.files.export.sample.TemplateService;
-import life.qbic.datamanager.importing.SampleInformationExtractor;
-import life.qbic.datamanager.importing.SampleInformationExtractor.SampleInformationForExistingSample;
-import life.qbic.datamanager.importing.parser.MetadataParser.ParsingException;
-import life.qbic.datamanager.importing.parser.ParsingResult;
-import life.qbic.datamanager.importing.parser.xlsx.XLSXParser;
+import life.qbic.datamanager.files.parsing.MetadataParser.ParsingException;
+import life.qbic.datamanager.files.parsing.ParsingResult;
+import life.qbic.datamanager.files.parsing.SampleInformationExtractor;
+import life.qbic.datamanager.files.parsing.SampleInformationExtractor.SampleInformationForExistingSample;
+import life.qbic.datamanager.files.parsing.xlsx.XLSXParser;
 import life.qbic.datamanager.views.general.WizardDialogWindow;
 import life.qbic.datamanager.views.general.upload.UploadWithDisplay;
 import life.qbic.datamanager.views.general.upload.UploadWithDisplay.FileType;
